Move cert_expiry base entity to separate module
